+++ changelog.md
@@ -1,3 +1,8 @@
+# 1.1.0
+- 2.2081 support
+- Fixed a visual glitch on the copy+ layer
+- Cleaned up code and made portions more readable
+
 # 1.0.12
 - Jitless iOS
 - Fixed copy+ resetting